+++ src/main/java/in/koreatech/koin/domain/ownershop/controller/OwnerShopApi.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import in.koreatech.koin.domain.ownershop.dto.CreateEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.ModifyEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.OwnerShopsRequest;
 import in.koreatech.koin.domain.ownershop.dto.OwnerShopsResponse;
-import in.koreatech.koin.domain.ownershop.dto.CreateEventRequest;
 import in.koreatech.koin.domain.shop.dto.CreateCategoryRequest;
 import in.koreatech.koin.domain.shop.dto.CreateMenuRequest;
 import in.koreatech.koin.domain.shop.dto.MenuCategoriesResponse;
@@ -49,7 +49,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @GetMapping("/owner/shops")
     ResponseEntity<OwnerShopsResponse> getOwnerShops(
-        @Auth(permit = {OWNER}) Long userId
+        @Auth(permit = {OWNER}) Integer userId
     );
 
     @ApiResponses(
@@ -64,7 +64,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @PostMapping("/owner/shops")
     ResponseEntity<Void> createOwnerShops(
-        @Auth(permit = {OWNER}) Long userId,
+        @Auth(permit = {OWNER}) Integer userId,
         @RequestBody @Valid OwnerShopsRequest ownerShopsRequest
     );
 
@@ -80,7 +80,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @GetMapping("/owner/shops/{id}")
     ResponseEntity<ShopResponse> getOwnerShopByShopId(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable Long id
     );
 
@@ -96,7 +96,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @GetMapping("/owner/shops/menus/{menuId}")
     ResponseEntity<MenuDetailResponse> getMenuByMenuId(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("menuId") Long id
     );
 
@@ -112,7 +112,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @GetMapping("/owner/shops/menus")
     ResponseEntity<ShopMenuResponse> getMenus(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @RequestParam("shopId") Long shopId
     );
 
@@ -128,7 +128,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @GetMapping("/owner/shops/menus/categories")
     ResponseEntity<MenuCategoriesResponse> getCategories(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @RequestParam("shopId") Long shopId
     );
 
@@ -144,7 +144,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @DeleteMapping("/owner/shops/menus/{menuId}")
     ResponseEntity<Void> deleteMenuByMenuId(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("menuId") Long menuId
     );
 
@@ -161,7 +161,7 @@ public interface OwnerShopApi {
     @SecurityRequirement(name = "Jwt Authentication")
     @DeleteMapping("/owner/shops/menus/categories/{categoryId}")
     ResponseEntity<Void> deleteCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("categoryId") Long categoryId
     );
 
@@ -177,7 +177,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점의 메뉴 생성")
     @PostMapping("/owner/shops/{id}/menus")
     ResponseEntity<Void> createMenu(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("id") Long shopId,
         @RequestBody @Valid CreateMenuRequest createMenuRequest
     );
@@ -194,7 +194,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점의 메뉴 카테고리 생성")
     @PostMapping("/owner/shops/{id}/menus/categories")
     ResponseEntity<Void> createMenuCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("id") Long shopId,
         @RequestBody @Valid CreateCategoryRequest createCategoryRequest
     );
@@ -211,7 +211,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점의 메뉴 수정")
     @PutMapping("/owner/shops/menus/{menuId}")
     ResponseEntity<Void> modifyMenu(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("menuId") Long menuId,
         @RequestBody @Valid ModifyMenuRequest modifyMenuRequest
     );
@@ -229,7 +229,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점의 메뉴 카테고리 수정")
     @PutMapping("/owner/shops/menus/categories/{categoryId}")
     ResponseEntity<Void> modifyMenuCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("categoryId") Long categoryId,
         @RequestBody @Valid ModifyCategoryRequest modifyCategoryRequest
     );
@@ -246,7 +246,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점 수정")
     @PutMapping("/owner/shops/{id}")
     ResponseEntity<Void> modifyOwnerShop(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("id") Long shopId,
         @RequestBody @Valid ModifyShopRequest modifyShopRequest
     );
@@ -263,7 +263,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점 이벤트 추가")
     @PostMapping("/owner/shops/{shopId}/event")
     ResponseEntity<Void> createShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("shopId") Long shopId,
         @RequestBody @Valid CreateEventRequest shopEventRequest
     );
@@ -281,7 +281,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점 이벤트 수정")
     @PutMapping("/owner/shops/{shopId}/event/{eventId}")
     ResponseEntity<Void> modifyShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("shopId") Long shopId,
         @PathVariable("eventId") Long eventId,
         @RequestBody @Valid ModifyEventRequest modifyEventRequest
@@ -298,7 +298,7 @@ public interface OwnerShopApi {
     @Operation(summary = "상점 이벤트 삭제")
     @DeleteMapping("/owner/shops/{shopId}/event/{eventId}")
     ResponseEntity<Void> deleteShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("shopId") Long shopId,
         @PathVariable("eventId") Long eventId
     );
